Rename the tests as .spec.ts files
