+++ package.json
@@ -57,7 +57,7 @@
   "ava": {
     "require": [
       "babel-register",
-      "./test/helpers/setup-browser-env.js"
+      "./test/setup-browser-env.js"
     ]
   },
   "dependencies": {
