--- test/coverage_check_test.ts
@@ -1,6 +1,0 @@
-import test from 'ava';
-import * as coveoanalytics from '../src/index';
-
-test('coverage', t => {
-    const _ = coveoanalytics;
-});
